Fix vietnamese transation
Using '&' is appealing, but the translated string is used in .xml files
and also in .desktop files. Gettext 0.25+ no longer escapes the translations
when merging them into xml files, which would mean we'd need 7amp; - that
in turn would not be parsed by the desktops and would result in the string
being shown literally.

Avoid the usage og '&' in this case